data models structure refactor
