+++ data/src/main/java/com/example/data/backend/ApiService.kt
@@ -2,6 +2,16 @@ package com.example.data.backend
 
 
 import com.example.data.backend.models.*
+import com.example.data.backend.models.requests.AuthRequest
+import com.example.data.backend.models.requests.RegistrationRequest
+import com.example.data.backend.models.requests.ResetCompleteRequest
+import com.example.data.backend.models.requests.SendEmailPasswordResetRequest
+import com.example.data.backend.models.requests.SendResetCodeRequest
+import com.example.data.backend.models.responses.LoginSuccess
+import com.example.data.backend.models.responses.RegistrationResponse
+import com.example.data.backend.models.responses.ResetCompleteResponse
+import com.example.data.backend.models.responses.SendCodeResponse
+import com.example.data.backend.models.responses.SendEmailPasswordResetSuccess
 import com.example.domain.utils.Endpoints
 import retrofit2.http.Body
 import retrofit2.http.POST
@@ -10,7 +20,7 @@ import retrofit2.http.POST
 interface  ApiService  {
 
     @POST(Endpoints.LOGIN_ENDPOINT)
-    suspend fun loginAuthorization(@Body authRequest: AuthRequest): LoginSucces
+    suspend fun loginAuthorization(@Body authRequest: AuthRequest): LoginSuccess
 
     @POST(Endpoints.SEND_EMAIL_PASSWORD_RESET_ENDPOINT)
     suspend fun sendEmailPasswordReset(@Body sendEmailPasswordResetRequest: SendEmailPasswordResetRequest) : SendEmailPasswordResetSuccess
