--- data/src/main/java/com/example/data/backend/models/Data.kt
@@ -1,8 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class Data(
-   @Json (name = "email") val email: String,
-   @Json (name = "tokens") val tokens: Tokens
-)
