--- data/src/main/java/com/example/data/backend/models/DataResetCompleteError.kt
@@ -1,8 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class DataResetCompleteError(
-    @Json (name = "errors") val errors: List<ResetCompleteErrors>,
-    @Json (name = "type") val type: String,
-)
