--- data/src/main/java/com/example/data/backend/models/DataResetCompleteResponse.kt
@@ -1,7 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class DataResetCompleteResponse(
-   @Json (name = "success") val success: String
-)
