--- data/src/main/java/com/example/data/backend/models/DataSendCode.kt
@@ -1,7 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class DataSendCode(
-    @Json (name = "success") val success: String
-)
