--- data/src/main/java/com/example/data/backend/models/EmailPassDataError.kt
@@ -1,8 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class EmailPassDataError(
-   @Json (name = "errors") val errors: List<EmailPassResetErrors>,
-   @Json (name = "type") val type: String
-)
