--- data/src/main/java/com/example/data/backend/models/EmailPassResetErrors.kt
@@ -1,7 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class EmailPassResetErrors(
-    @Json (name = "detail") val detail: String
-)
