--- data/src/main/java/com/example/data/backend/models/LoginErrors.kt
@@ -1,7 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class LoginErrors(
-    @Json(name = "detail") val detail: String
-)
