--- data/src/main/java/com/example/data/backend/models/SendCodeDataError.kt
@@ -1,8 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class SendCodeDataError(
-   @Json (name = "errors") val errors: List<SendCodeErrors>,
-   @Json (name = "type") val type: String
-)
