--- data/src/main/java/com/example/data/backend/models/SendCodeError.kt
@@ -1,10 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class SendCodeError(
-    @Json (name = "code")val code: Int,
-    @Json (name = "`data")val `data`: SendCodeDataError,
-    @Json (name = "message")val message: Any? = null,
-    @Json (name = "status")val status: String
-)
