--- data/src/main/java/com/example/data/backend/models/SendCodeErrors.kt
@@ -1,7 +1,0 @@
-package com.example.data.backend.models
-
-import com.squareup.moshi.Json
-
-data class SendCodeErrors(
-   @Json (name = "detail") val detail: String
-)
