+++ data/src/main/java/com/example/data/backend/models/requests/SendResetCodeRequest.kt
@@ -1,4 +1,4 @@
-package com.example.data.backend.models
+package com.example.data.backend.models.requests
 
 import com.squareup.moshi.Json
