+++ data/src/main/java/com/example/data/backend/models/responses/LoginError.kt
@@ -1,4 +1,4 @@
-package com.example.data.backend.models
+package com.example.data.backend.models.responses
 
 import com.squareup.moshi.Json
 
@@ -7,4 +7,13 @@ data class LoginError(
    @Json (name = "`data`") val `data`: DataError,
    @Json (name = "message") val message: Any? = null,
    @Json (name = "status") val status: String
+)
+
+data class DataError(
+   @Json(name = "errors") val errors: List<LoginErrors>,
+   @Json(name = "type") val type: String
+)
+
+data class LoginErrors(
+    @Json(name = "detail") val detail: String
 )
