+++ data/src/main/java/com/example/data/backend/models/responses/SendEmailPasswordResetSuccess.kt
@@ -1,4 +1,4 @@
-package com.example.data.backend.models
+package com.example.data.backend.models.responses
 
 import com.squareup.moshi.Json
 
@@ -7,4 +7,8 @@ data class SendEmailPasswordResetSuccess (
     @Json (name = "`data`") val `data`: DataEmailReset,
     @Json (name = "message") val message: Any? = null,
     @Json (name = "status") val status: String
+)
+
+data class DataEmailReset(
+@Json(name = "success") val success: String
 )
